PacbioLibraryCreate modification
- Composition api
- Using new Pinia store
- Unit test case update
